requirements.txt
bleach==1.4.3
cffi==1.7.0
cryptography==1.4
Django==1.10
django-appconf==1.0.2
django-compressor==2.1
django-libsass==0.7
djangorestframework==3.4.5
enum34==1.1.6
gunicorn==19.6.0
html5lib==0.9999999
idna==2.1
ipaddress==1.0.16
libsass==0.11.1
Markdown==2.6.6
Pillow==3.3.1
pyapns==0.4.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
Twisted==16.3.2
Unipath==1.1
zope.interface==4.2.0
